+++ jme3-materialeditor/src/com/jme3/gde/materialdefinition/editor/previews/BasePreview.java
@@ -1,7 +1,3 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package com.jme3.gde.materialdefinition.editor.previews;
 
 import com.jme3.gde.materials.MaterialProperty;
